refactor: extract DebugEntryList shared component, add layout-aware search
- Extract duplicated debug entry list from IndexPage and ListPage into
  shared DebugEntryList component
- Use searchVariants() from layoutTranslit for keyboard layout-aware
  filtering (QWERTY/ЙЦУКЕН), same as CommandPalette
- Search by URL, method, status code, command, or ID
- useDeferredValue for smooth typing during filtering
- Full-width search bar with clear button and count badge
