+++ libs/frontend/packages/panel/src/Application/Pages/IndexPage.tsx
@@ -1,33 +1,16 @@
+import {DebugEntryList} from '@app-dev-panel/panel/Module/Debug/Component/DebugEntryList';
 import {useLazyGetGeneratorsQuery} from '@app-dev-panel/panel/Module/GenCode/API/GenCode';
 import {useLazyGetParametersQuery} from '@app-dev-panel/panel/Module/Inspector/API/Inspector';
 import {useSelector} from '@app-dev-panel/panel/store';
 import {addFavoriteUrl, changeBaseUrl, removeFavoriteUrl} from '@app-dev-panel/sdk/API/Application/ApplicationContext';
-import {changeEntryAction} from '@app-dev-panel/sdk/API/Debug/Context';
-import {DebugEntry, useGetDebugQuery} from '@app-dev-panel/sdk/API/Debug/Debug';
-import {FilterInput} from '@app-dev-panel/sdk/Component/FilterInput';
+import {useGetDebugQuery} from '@app-dev-panel/sdk/API/Debug/Debug';
 import {PageHeader} from '@app-dev-panel/sdk/Component/PageHeader';
 import {StatusCard} from '@app-dev-panel/sdk/Component/StatusCard';
 import {primitives} from '@app-dev-panel/sdk/Component/Theme/tokens';
-import {isDebugEntryAboutConsole, isDebugEntryAboutWeb} from '@app-dev-panel/sdk/Helper/debugEntry';
-import {formatBytes} from '@app-dev-panel/sdk/Helper/formatBytes';
-import {formatDate} from '@app-dev-panel/sdk/Helper/formatDate';
-import {
-    Alert,
-    AlertTitle,
-    Box,
-    Chip,
-    CircularProgress,
-    Icon,
-    IconButton,
-    InputBase,
-    Tooltip,
-    Typography,
-    type Theme,
-} from '@mui/material';
-import {styled, useTheme} from '@mui/material/styles';
-import {useCallback, useEffect, useMemo, useState} from 'react';
+import {Chip, Icon, IconButton, InputBase, Typography} from '@mui/material';
+import {styled} from '@mui/material/styles';
+import {useEffect, useState} from 'react';
 import {useDispatch} from 'react-redux';
-import {useNavigate} from 'react-router-dom';
 
 // ---------------------------------------------------------------------------
 // Styled components
@@ -83,81 +66,6 @@ const SectionLabel = styled(Typography)(({theme}) => ({
     borderBottom: `1px solid ${theme.palette.divider}`,
 }));
 
-// --- Debug list styled components ---
-
-const EntryRow = styled(Box)(({theme}) => ({
-    display: 'flex',
-    alignItems: 'center',
-    gap: theme.spacing(1.5),
-    padding: theme.spacing(1, 1.5),
-    borderBottom: `1px solid ${theme.palette.divider}`,
-    cursor: 'pointer',
-    transition: 'background-color 0.1s ease',
-    '&:hover': {backgroundColor: theme.palette.action.hover},
-}));
-
-const MethodLabel = styled(Typography)({
-    fontFamily: primitives.fontFamilyMono,
-    fontSize: '11px',
-    fontWeight: 600,
-    minWidth: 44,
-    flexShrink: 0,
-});
-
-const PathLabel = styled(Typography)({
-    fontFamily: primitives.fontFamilyMono,
-    fontSize: '13px',
-    flex: 1,
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    whiteSpace: 'nowrap',
-});
-
-const MetaLabel = styled(Typography)({fontFamily: primitives.fontFamilyMono, fontSize: '10px', flexShrink: 0});
-
-const StatusChip = styled(Chip)({fontSize: '10px', height: 20, minWidth: 36, fontWeight: 600, borderRadius: 4});
-
-const StatCell = styled(Box)(({theme}) => ({
-    display: 'flex',
-    alignItems: 'center',
-    gap: theme.spacing(0.5),
-    flexShrink: 0,
-}));
-
-const StatLabel = styled(Typography)({fontFamily: primitives.fontFamilyMono, fontSize: '10px'});
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-const methodColor = (method: string, theme: Theme): string => {
-    switch (method.toUpperCase()) {
-        case 'GET':
-            return theme.palette.success.main;
-        case 'POST':
-            return theme.palette.primary.main;
-        case 'PUT':
-        case 'PATCH':
-            return theme.palette.warning.main;
-        case 'DELETE':
-            return theme.palette.error.main;
-        default:
-            return theme.palette.text.secondary;
-    }
-};
-
-const statusColor = (status: number, theme: Theme): string => {
-    if (status >= 500) return theme.palette.error.main;
-    if (status >= 400) return theme.palette.warning.main;
-    return theme.palette.success.main;
-};
-
-const statusBg = (status: number, theme: Theme): string => {
-    if (status >= 500) return theme.palette.error.light;
-    if (status >= 400) return theme.palette.primary.light;
-    return theme.palette.primary.light;
-};
-
 // ---------------------------------------------------------------------------
 // Component
 // ---------------------------------------------------------------------------
@@ -165,8 +73,6 @@ const statusBg = (status: number, theme: Theme): string => {
 export function IndexPage() {
     const defaultBackendUrl = useSelector((state) => state.application.baseUrl) as string;
     const dispatch = useDispatch();
-    const theme = useTheme();
-    const navigate = useNavigate();
     const [inspectorQuery] = useLazyGetParametersQuery();
     const [genCodeQuery] = useLazyGetGeneratorsQuery();
     const baseUrl = useSelector((state) => state.application.baseUrl);
@@ -178,16 +84,13 @@ export function IndexPage() {
     });
     const favoriteUrls = useSelector((state) => state.application.favoriteUrls) as string[];
 
-    // Debug list state — also drives debug status card (avoids race with Layout's resetApiState)
+    // Debug query — used for status card
     const {
-        data: entries,
-        isLoading: debugLoading,
         isFetching: debugFetching,
         isSuccess: debugSuccess,
         isError: debugError,
         refetch: debugRefetch,
     } = useGetDebugQuery();
-    const [filter, setFilter] = useState('');
 
     // Derive debug status from the shared useGetDebugQuery hook (managed by Layout)
     const debugStatus: 'connected' | 'disconnected' | 'loading' = debugFetching
@@ -219,25 +122,6 @@ export function IndexPage() {
         event.preventDefault();
         await handleChangeUrl(url);
     };
-
-    const handleEntryClick = useCallback(
-        (entry: DebugEntry) => {
-            dispatch(changeEntryAction(entry));
-            navigate(`/debug?debugEntry=${entry.id}`);
-        },
-        [dispatch, navigate],
-    );
-
-    const filtered = useMemo(() => {
-        if (!entries) return [];
-        if (!filter.trim()) return entries;
-        const q = filter.toLowerCase();
-        return entries.filter((entry) => {
-            const path = entry.request?.path ?? entry.command?.input ?? '';
-            const method = entry.request?.method ?? '';
-            return path.toLowerCase().includes(q) || method.toLowerCase().includes(q) || entry.id.includes(q);
-        });
-    }, [entries, filter]);
 
     useEffect(() => {
         checkStatus();
@@ -332,132 +216,7 @@ export function IndexPage() {
             )}
 
             <SectionLabel>Debug Entries</SectionLabel>
-            {debugLoading ? (
-                <Box sx={{display: 'flex', justifyContent: 'center', py: 4}}>
-                    <CircularProgress size={24} />
-                </Box>
-            ) : !entries || entries.length === 0 ? (
-                <Alert severity="info" sx={{mb: 2}}>
-                    <AlertTitle>No debug entries</AlertTitle>
-                    Make sure the debugger is enabled and your application has processed requests.
-                </Alert>
-            ) : (
-                <Box>
-                    <Box sx={{display: 'flex', alignItems: 'center', gap: 1, mb: 1}}>
-                        <FilterInput value={filter} onChange={setFilter} placeholder="Filter entries..." />
-                        <Typography sx={{fontSize: '11px', color: 'text.disabled', ml: 0.5}}>
-                            {filtered.length} / {entries.length}
-                        </Typography>
-                        <Box sx={{flex: 1}} />
-                        <Tooltip title={debugFetching ? 'Refreshing...' : 'Refresh'}>
-                            <IconButton size="small" onClick={() => debugRefetch()} disabled={debugFetching}>
-                                <Icon sx={{fontSize: 18}}>{debugFetching ? 'hourglass_empty' : 'refresh'}</Icon>
-                            </IconButton>
-                        </Tooltip>
-                    </Box>
-                    {filtered.map((entry) => {
-                        if (isDebugEntryAboutWeb(entry)) {
-                            const duration = entry.web?.request?.processingTime;
-                            const memory = entry.web?.memory?.peakUsage;
-                            return (
-                                <EntryRow key={entry.id} onClick={() => handleEntryClick(entry)}>
-                                    <MethodLabel sx={{color: methodColor(entry.request.method, theme)}}>
-                                        {entry.request.method}
-                                    </MethodLabel>
-                                    <PathLabel>{entry.request.path}</PathLabel>
-                                    {duration != null && (
-                                        <MetaLabel sx={{color: 'primary.main'}}>
-                                            {(duration * 1000).toFixed(0)}ms
-                                        </MetaLabel>
-                                    )}
-                                    {memory != null && (
-                                        <MetaLabel sx={{color: 'success.main'}}>{formatBytes(memory)}</MetaLabel>
-                                    )}
-                                    {entry.logger?.total != null && entry.logger.total > 0 && (
-                                        <StatCell>
-                                            <Icon sx={{fontSize: 13, color: 'text.disabled'}}>description</Icon>
-                                            <StatLabel sx={{color: 'text.disabled'}}>{entry.logger.total}</StatLabel>
-                                        </StatCell>
-                                    )}
-                                    {entry.db?.queries?.total != null && entry.db.queries.total > 0 && (
-                                        <StatCell>
-                                            <Icon
-                                                sx={{
-                                                    fontSize: 13,
-                                                    color: entry.db.queries.error ? 'error.main' : 'text.disabled',
-                                                }}
-                                            >
-                                                storage
-                                            </Icon>
-                                            <StatLabel
-                                                sx={{color: entry.db.queries.error ? 'error.main' : 'text.disabled'}}
-                                            >
-                                                {entry.db.queries.total}
-                                            </StatLabel>
-                                        </StatCell>
-                                    )}
-                                    {entry.exception && (
-                                        <Tooltip title={entry.exception.message}>
-                                            <Icon sx={{fontSize: 15, color: 'error.main'}}>error</Icon>
-                                        </Tooltip>
-                                    )}
-                                    <StatusChip
-                                        label={entry.response.statusCode}
-                                        size="small"
-                                        sx={{
-                                            color: statusColor(entry.response.statusCode, theme),
-                                            backgroundColor: statusBg(entry.response.statusCode, theme),
-                                        }}
-                                    />
-                                    <MetaLabel sx={{color: 'text.disabled', minWidth: 55}}>
-                                        {formatDate(entry.web.request.startTime)}
-                                    </MetaLabel>
-                                </EntryRow>
-                            );
-                        }
-
-                        if (isDebugEntryAboutConsole(entry)) {
-                            const duration = entry.console?.request?.processingTime;
-                            const memory = entry.console?.memory?.peakUsage;
-                            const exitOk = entry.command?.exitCode === 0;
-                            return (
-                                <EntryRow key={entry.id} onClick={() => handleEntryClick(entry)}>
-                                    <Icon sx={{fontSize: 16, color: 'text.disabled', flexShrink: 0}}>terminal</Icon>
-                                    <PathLabel>{entry.command?.input ?? 'Unknown command'}</PathLabel>
-                                    {duration != null && (
-                                        <MetaLabel sx={{color: 'primary.main'}}>
-                                            {(duration * 1000).toFixed(0)}ms
-                                        </MetaLabel>
-                                    )}
-                                    {memory != null && (
-                                        <MetaLabel sx={{color: 'success.main'}}>{formatBytes(memory)}</MetaLabel>
-                                    )}
-                                    {entry.logger?.total != null && entry.logger.total > 0 && (
-                                        <StatCell>
-                                            <Icon sx={{fontSize: 13, color: 'text.disabled'}}>description</Icon>
-                                            <StatLabel sx={{color: 'text.disabled'}}>{entry.logger.total}</StatLabel>
-                                        </StatCell>
-                                    )}
-                                    <StatusChip
-                                        label={exitOk ? 'OK' : `EXIT ${entry.command?.exitCode}`}
-                                        size="small"
-                                        color={exitOk ? 'success' : 'error'}
-                                    />
-                                    <MetaLabel sx={{color: 'text.disabled', minWidth: 55}}>
-                                        {formatDate(entry.console.request.startTime)}
-                                    </MetaLabel>
-                                </EntryRow>
-                            );
-                        }
-
-                        return (
-                            <EntryRow key={entry.id} onClick={() => handleEntryClick(entry)}>
-                                <PathLabel>{entry.id}</PathLabel>
-                            </EntryRow>
-                        );
-                    })}
-                </Box>
-            )}
+            <DebugEntryList />
         </>
     );
 }
